+++ internal/vehicleoccupancies/contextOditi_test.go
@@ -74,6 +74,7 @@ func TestStopPointFileWithOutField(t *testing.T) {
 	spFileName = "mapping_stops0.csv"
 
 	vehicleOccupanciesOditiContext := &VehicleOccupanciesOditiContext{}
+	vehicleOccupanciesOditiContext.voContext = &VehicleOccupanciesContext{}
 
 	uri, err := url.Parse(fmt.Sprintf("file://%s/", fixtureDir))
 	require.Nil(err)
@@ -105,6 +106,7 @@ func TestCoursesFileWithOutField(t *testing.T) {
 	courseFileName = "extraction_courses0.csv"
 
 	vehicleOccupanciesOditiContext := &VehicleOccupanciesOditiContext{}
+	vehicleOccupanciesOditiContext.voContext = &VehicleOccupanciesContext{}
 
 	uri, err := url.Parse(fmt.Sprintf("file://%s/", fixtureDir))
 	require.Nil(err)
@@ -174,6 +176,7 @@ func TestDataManagerForVehicleOccupancies(t *testing.T) {
 	require.Nil(err)
 
 	vehicleOccupanciesOditiContext := &VehicleOccupanciesOditiContext{}
+	vehicleOccupanciesOditiContext.voContext = &VehicleOccupanciesContext{}
 	vehiculeOccupanciesContext := &VehicleOccupanciesContext{}
 
 	// Load StopPoints
@@ -403,6 +406,7 @@ func TestStatusForVehicleOccupancies(t *testing.T) {
 	require.Nil(err)
 
 	vehicleOccupanciesOditiContext := &VehicleOccupanciesOditiContext{}
+	vehicleOccupanciesOditiContext.voContext = &VehicleOccupanciesContext{}
 	vehiculeOccupanciesContext := &VehicleOccupanciesContext{}
 
 	// Load StopPoints
